update: modify usestate to be borrowed
